Added files via upload
This is the updated game.py file that works. It already has the Brain class which acts as the neural net for the Agents. I implemented DEAP with 100 generations, and I added a modifiedEASimple method that resets the agents list every generation and undergoes the calculations for fitness, while still having the exact same lines as the source code for everything else. When run it should do the first generation in pygame but the subsequent calculations are not shown. The final generation is shown to see progress.