Added files via upload
Changed the sigmoid function to 2/(1+e^{-x}) so that the range of the output covers the interval (0,2); I did this cause the config.py file has a default track speed of 2 for each direction. This actually increased the fitnesses. In addition, I changed the crossover to just using uniform crossover, because I noticed that the blend crossover was removing any good individuals that popped out per generation (which caused some of the downgrading of fitnesses). I tweaked the mutation function by only adding -.5 to .5 instead of -1 to 1. I cleaned up the genetic algorithm so it wasn't completely the original source code. I also added extra parts for matplotlib which shouldn't be needed for the final result. The main program goes through each number of neurons but we can change it to only testing for 5 since that was shown to be the best~ Also, I went through the neural network and I found a mistake with the bias, but I fixed that so it should be good. An important thing I changed though was